Harden Grok Responses path and fix protocol clippy CI
Address review findings for PR #645: split the Responses client, default
store=false with XAI_STORE opt-in, price Grok in architect accounting,
normalize OpenAI-shaped tools, and unit-test production SSE helpers.
Remove a useless format! borrow that failed clippy on protocol/cli/ui jobs.
